AFT: Move cutter.py to cutters/ and device.py to devices/ folder
Having cutter.py in cutters/ folder and device.py in devices/ folder
makes more sense than having the base class files in the main AFT
folder.
