+++ cutters/ethernetrelay16.py
@@ -65,7 +65,7 @@ dec   hex
 """
 
 
-from aft.cutter import Cutter
+from aft.cutters.cutter import Cutter
 
 class EthernetRelay16(Cutter):
     """
