+++ devices/beagleboneblackdevice.py
@@ -30,7 +30,7 @@ except ImportError:
     import subprocess as subprocess32
 
 from aft.logger import Logger as logger
-from aft.device import Device
+from aft.devices.device import Device
 import aft.config as config
 import aft.errors as errors
 import aft.tools.ssh as ssh
